Added some error handeling
