--- SupplierForm/Form1.cs
@@ -1,10 +1,0 @@
-namespace SupplierForm
-{
-    public partial class Form1 : Form
-    {
-        public Form1()
-        {
-            InitializeComponent();
-        }
-    }
-}
