--- WinFormsApp1/Form1.cs
@@ -1,10 +1,0 @@
-namespace WinFormsApp1
-{
-    public partial class Form1 : Form
-    {
-        public Form1()
-        {
-            InitializeComponent();
-        }
-    }
-}
