--- inventryUI/Views/ISupplierView.cs
@@ -1,11 +1,0 @@
-﻿using inventryUI.Models;
-using System.Collections.Generic;
-
-namespace inventryUI.Views
-{
-    public interface ISupplierView
-    {
-        void DisplaySuppliers(List<Supplier> suppliers);
-        Supplier GetSelectedSupplier();
-    }
-}
